+++ src/main/java/com/moabam/api/application/AuthenticationService.java
@@ -14,7 +14,7 @@ import com.moabam.api.dto.AuthorizationTokenInfoResponse;
 import com.moabam.api.dto.AuthorizationTokenRequest;
 import com.moabam.api.dto.AuthorizationTokenResponse;
 import com.moabam.api.dto.OAuthMapper;
-import com.moabam.global.common.constant.GlobalConstant;
+import com.moabam.global.common.util.GlobalConstant;
 import com.moabam.global.common.util.TokenConstant;
 import com.moabam.global.config.OAuthConfig;
 import com.moabam.global.error.exception.BadRequestException;
