+++ src/test/java/com/moabam/api/application/OAuth2AuthorizationServerRequestServiceTest.java
@@ -31,7 +31,7 @@ import org.springframework.web.client.RestTemplate;
 
 import com.moabam.api.dto.AuthorizationTokenInfoResponse;
 import com.moabam.api.dto.AuthorizationTokenResponse;
-import com.moabam.global.common.constant.GlobalConstant;
+import com.moabam.global.common.util.GlobalConstant;
 import com.moabam.global.error.exception.BadRequestException;
 import com.moabam.global.error.model.ErrorMessage;
