+++ src/test/java/com/moabam/api/presentation/MemberControllerTest.java
@@ -38,7 +38,7 @@ import com.moabam.api.application.AuthenticationService;
 import com.moabam.api.application.OAuth2AuthorizationServerRequestService;
 import com.moabam.api.dto.AuthorizationCodeResponse;
 import com.moabam.api.dto.AuthorizationTokenResponse;
-import com.moabam.global.common.constant.GlobalConstant;
+import com.moabam.global.common.util.GlobalConstant;
 import com.moabam.global.config.OAuthConfig;
 import com.moabam.global.error.handler.RestTemplateResponseHandler;
 import com.moabam.support.fixture.AuthorizationResponseFixture;
