+++ src/test/java/com/moabam/support/fixture/ParticipantFixture.java
@@ -1,4 +1,4 @@
-package com.moabam.fixture;
+package com.moabam.support.fixture;
 
 import com.moabam.api.domain.entity.Participant;
 import com.moabam.api.domain.entity.Room;
