+++ src/test/java/com/moabam/support/fixture/RoomFixture.java
@@ -1,4 +1,4 @@
-package com.moabam.fixture;
+package com.moabam.support.fixture;
 
 import com.moabam.api.domain.entity.Room;
 import com.moabam.api.domain.entity.enums.RoomType;
